refactor(utilities): move utilities into a single file
as well as a minor cleanup of the lifecycleInvoke method
