--- src/lifecycle.js
@@ -1,17 +1,0 @@
-export function invokeLifecycle(instance: any, name: string, model: any) {
-  if (typeof instance[name] === 'function') {
-    let result = instance[name](model);
-
-    if (result instanceof Promise) {
-      return result;
-    }
-
-    if (result !== null && result !== undefined) {
-      return Promise.resolve(result);
-    }
-
-    return Promise.resolve(true);
-  }
-
-  return Promise.resolve(true);
-}
